requirements.txt
aiomqtt
aiohttp
pyyaml